voice bot/sih_pritam/hybrid_live_interviewer/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
httpx==0.24.1
python-dotenv==1.0.0
websockets==11.0.3
aiofiles==23.1.0
typing-extensions==4.7.1
redis==4.5.5
aiocache==0.12.5
soundfile==0.12.1
faster-whisper==0.8.6
TTS==0.14.0
langgraph==0.0.0
pytest==7.4.0
